backend/pyproject.toml
[project]
name = "unmapped-backend"
version = "0.1.0"
description = "UNMAPPED — open skills infrastructure backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "pydantic>=2.9",
    "pydantic-settings>=2.5",
    "duckdb>=1.1",
    "pyyaml>=6.0",
    "python-multipart>=0.0.12",
    "openai>=1.50",
    "httpx>=0.27",
    "pandas>=2.2",
    "qrcode[pil]>=7.4",
    "python-dotenv>=1.0",
    "pypdf>=5.0",
    "python-docx>=1.1",
]

[project.optional-dependencies]
dev = ["ruff>=0.6", "pytest>=8.3"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]
